requirements.txt
absl-py @ file:///C:/ci/absl-py_1623867344508/work
aiohttp @ file:///C:/ci/aiohttp_1614361236131/work
astor==0.8.1
astunparse==1.6.3
async-timeout==3.0.1
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
blinker==1.4
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
certifi==2021.5.30
cffi @ file:///C:/ci/cffi_1625831763871/work
chardet @ file:///C:/ci/chardet_1605285153095/work
click @ file:///tmp/build/80754af9/click_1621604852318/work
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
coverage @ file:///C:/ci/coverage_1614614921360/work
cryptography @ file:///C:/ci/cryptography_1616769514516/work
cycler==0.10.0
Cython @ file:///C:/ci/cython_1626257069493/work
decorator @ file:///tmp/build/80754af9/decorator_1621259047763/work
Flask @ file:///home/ktietz/src/ci/flask_1611932660458/work
flatbuffers==1.12
gast==0.3.3
google-auth @ file:///tmp/build/80754af9/google-auth_1626320605116/work
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1617120569401/work
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.1.0
h5py==2.10.0
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
idna-ssl @ file:///C:/ci/idna_ssl_1611752657819/work
importlib-metadata @ file:///C:/ci/importlib-metadata_1617877476772/work
ipykernel @ file:///C:/ci/ipykernel_1596206775157/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1593446240034/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
itsdangerous @ file:///tmp/build/80754af9/itsdangerous_1621432558163/work
jedi==0.17.0
Jinja2 @ file:///tmp/build/80754af9/jinja2_1624781299557/work
joblib @ file:///tmp/build/80754af9/joblib_1613502643832/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1616770841739/work
jupyter-core @ file:///C:/ci/jupyter_core_1612213536848/work
Keras==2.4.3
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
kiwisolver @ file:///C:/ci/kiwisolver_1612282446297/work
Markdown @ file:///C:/ci/markdown_1614364003005/work
MarkupSafe @ file:///C:/ci/markupsafe_1621528313524/work
matplotlib @ file:///C:/ci/matplotlib-suite_1613408055530/work
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict @ file:///C:/ci/multidict_1607362067746/work
nltk @ file:///tmp/build/80754af9/nltk_1621347441292/work
numpy==1.19.5
oauthlib @ file:///tmp/build/80754af9/oauthlib_1623060228408/work
olefile==0.46
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
pandas @ file:///C:/ci/pandas_1608056614942/work
parso @ file:///tmp/build/80754af9/parso_1617223946239/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///C:/ci/pillow_1625663293114/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1616415428029/work
protobuf==3.17.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments @ file:///tmp/build/80754af9/pygments_1621606182707/work
PyJWT @ file:///C:/ci/pyjwt_1619651879817/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
pyreadline==2.1
PySocks @ file:///C:/ci/pysocks_1605305839978/work
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
pywin32==227
PyYAML==5.4.1
pyzmq==20.0.0
regex @ file:///C:/ci/regex_1625610189614/work
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scikit-learn @ file:///C:/ci/scikit-learn_1622739439730/work
scipy @ file:///C:/ci/scipy_1597675683670/work
six==1.15.0
tensorboard @ file:///home/builder/ktietz/aggregate/tensorflow_recipes/ci_te/tensorboard_1614593728657/work/tmp_pip_dir
tensorboard-plugin-wit==1.6.0
tensorflow==2.0.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
termcolor==1.1.0
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
tornado @ file:///C:/ci/tornado_1606942379977/work
tqdm @ file:///tmp/build/80754af9/tqdm_1625563689033/work
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3 @ file:///tmp/build/80754af9/urllib3_1625084269274/work
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
Werkzeug==0.16.1
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306197271/work
wincertstore==0.2
wrapt==1.12.1
yarl @ file:///C:/ci/yarl_1606921989557/work
zipp @ file:///tmp/build/80754af9/zipp_1625570634446/work
